Massive Homework/Classwork Submission
Because I had trouble with my environment for a good part of the
quarter, I had to redo a lot of assignments which I left in my local
memory that got lost. I really learned my lesson - that when working
with the REPL you also should be tracking progress in the IDE in R.
Sometimes I would just work the data until I had the thing I wanted, but
then my log of history would get wiped out and I would have to redo some
of the work.