538Assignment
I was unable to get the plot to show up correctly at the end. I even
tried manually crunching the numbers to create a simpler table to deal
with the mosaic function. Even after converting my data.frame to a
matrix to eliminate all forms of lists inside the structure; there were
still dimension issues in the mosaic call. Ultimately I made an
association graph which correctly displays the area of the factors.